+++ Hawk/changelog.md
@@ -98,4 +98,5 @@
 - Added search of mail items accessed to the User Investigation (Get-HawkUserMailItemsAccessed)
 - Added search of Exchange Search Queries to the User Investigation (Get-HawkUserExchangeSearchQuery)
 - Implemented check to verify that an Exchange operation is enabled for auditing before attempting to pull logs
+- Added log pull of user Send activity to the User Investigation (Get-HawkUserMailSendActivity)